+++ lib/widgets/charts/dashboard_health_chart.dart
@@ -184,11 +184,12 @@ class HealthChartInfoWidget extends StatelessWidget {
 
         return Positioned(
           top: 0,
-          left: 10,
+          left: 20,
           child: SizedBox(
             width: max(MediaQuery.of(context).size.width, 300) - 20,
             child: IgnorePointer(
               child: Row(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
                 children: [
                   Text(
                     healthType?.displayName ?? chartConfig.healthType,