+++ lib/widgets/charts/dashboard_measurables_line_chart.dart
@@ -145,7 +145,7 @@ class MeasurablesChartInfoWidget extends StatelessWidget {
 
         return Positioned(
           top: 0,
-          left: 10,
+          left: 20,
           child: SizedBox(
             width: max(MediaQuery.of(context).size.width, 300) - 20,
             child: Row(