+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Task duration estimate bottom sheet
 - Font colors on survey bottom sheet
 - Color issues
+- Charts layout
 
 ## [0.8.366] - 2023-06-03
 ### Changed: