+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Charts layout
 - Colors in dashboard chart selector
 - Colors in entry delete bottom sheet
+- Colors in measurables definition page
 
 ## [0.8.366] - 2023-06-03
 ### Changed: