+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added:
+- Config flag for following system brightness
+
+## [0.8.367] - 2023-06-05
 ### Changed:
 - Color schemes from flex_color_scheme
 - Styling